+++ src/rsrs.rs
@@ -1,9 +1,11 @@
 //! An implementation of the RSRS algorithm.
 
+pub mod args;
 pub mod box_skeletonisation;
 pub mod rsrs_cycle;
 pub mod rsrs_factors;
 pub mod sketch;
+pub mod statistics;
 pub mod tree_indexing;
 
 #[cfg(test)]
