+++ src/rsrs/rsrs_cycle.rs
@@ -3,24 +3,24 @@ use super::{
     sketch::SketchData,
     tree_indexing::{TreeData, TreeIndexing},
 };
-use crate::rsrs::{
-    rsrs_factors::{
-        commutative_factors::{
-            BoxType, CommutativeFactors, CommutativeFactorsOperations, DiagBoxFactor, Factor,
-            LevelIdFactors, RsrsFactors,
+use crate::{
+    rsrs::{
+        args::{RankPicking, RsrsOptions},
+        rsrs_factors::{
+            commutative_factors::{
+                BoxType, CommutativeFactors, CommutativeFactorsOperations, DiagBoxFactor, Factor,
+                MultiLevelIdFactors, RsrsFactors,
+            },
+            rsrs_operator::{FactType, LocalFromSpaces, RsrsFactorsImpl, RsrsOperator},
         },
-        null_and_extract::{ExtractOptions, IdOptions, PivotMethod},
-        rsrs_operator::{FactType, LocalFromSpaces, RsrsFactorsImpl, RsrsOperator},
+        sketch::{SamplingSpace, UpdateType},
         statistics::{
-            IdTimes, IdTimesOperations, LevelEffort, LuTimes, LuTimesOperations, Times,
-            UpdateTimes, UpdateTimesOperations,
+            IdTimes, IdTimesOperations, LevelEffort, LimitingFactors, LimitingLevel, LuTimes,
+            LuTimesOperations, Stats, Times, UpdateTimes, UpdateTimesOperations,
         },
     },
-    sketch::SamplingSpace,
+    utils::io::IOData,
 };
-use crate::utils::linear_algebra::NullMethod;
-use crate::{rsrs::sketch::Shift, utils::io::IOData};
-use crate::{rsrs::sketch::UpdateType, utils::linear_algebra::BlockExtractionMethod};
 use bempp_octree::{MortonKey, Octree};
 use mpi::traits::CommunicatorCollectives;
 use rand_distr::{Distribution, Standard, StandardNormal};
@@ -31,56 +31,15 @@ use rayon::{
 use rlst::dense::{linalg::lu::MatrixLu, tools::RandScalar};
 pub use rlst::prelude::*;
 use rustc_hash::FxHashSet;
-use serde::Deserialize;
 use std::{
     collections::HashMap,
-    fmt::Write,
     path::Path,
     time::{Duration, Instant},
-}; // Ensure IndexableSpace is in scope
+};
+
 type Inds<T> = Vec<Vec<T>>;
 
-#[derive(Debug)]
-pub struct LimitingLevel {
-    pub level: usize,
-    pub num_boxes: usize,
-    pub active_points: usize,
-    pub elapsed_time: u128,
-}
-
-#[derive(Debug)]
-pub struct LimitingFactors {
-    pub min_samples: usize,
-    pub max_level: usize,
-    pub limiting_level: LimitingLevel,
-    pub leaf_count: usize,
-}
-
-#[derive(Debug)]
-pub struct Stats {
-    pub sampling_time: Vec<u128>,
-    pub sampling_extraction_time: u128,
-    pub id_times: Vec<IdTimes>,
-    pub tot_id_time: u128,
-    pub lu_times: Vec<LuTimes>,
-    pub tot_lu_time: u128,
-    pub update_times: Vec<UpdateTimes>,
-    pub total_elapsed_time: u128,
-    pub total_elapsed_time_wo_sampling: u128,
-    pub dim: usize,
-    pub extraction_time: u128,
-    pub residual_size: usize,
-    pub ranks: Vec<usize>,
-    pub box_sizes: Vec<usize>,
-    pub near_field_sizes: Vec<usize>,
-    pub dec_boxes_per_level: Vec<usize>,
-    pub index_calculation: u128,
-    pub sorting_near_field: u128,
-    pub residual_calculation: u128,
-    pub limiting_factors: LimitingFactors,
-    pub level_effort: Vec<LevelEffort>,
-    pub mv_avg_time: Vec<u128>,
-}
+type Real<T> = <T as rlst::RlstScalar>::Real;
 
 pub struct Rsrs<Item: RlstScalar> {
     level_indexing: TreeData,
@@ -96,272 +55,6 @@ pub struct Rsrs<Item: RlstScalar> {
     pub stats: Stats,
     options: RsrsOptions<Item>,
 }
-
-#[derive(Debug, Clone, Deserialize)]
-pub enum RankPicking {
-    Min,
-    DoubleMin,
-    Max,
-    Avg,
-    Mid,
-    Tol,
-}
-
-#[derive(Debug, Clone)]
-pub struct SketchingOptions {
-    pub oversampling: usize,
-    pub oversampling_diag_blocks: usize,
-    pub initial_num_samples: usize,
-    pub min_num_samples: usize,
-    pub shift: Shift,
-    pub save_samples: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct RsrsOptions<Item: RlstScalar> {
-    pub fact_type: FactType,
-    pub sketching: SketchingOptions,
-    pub id_options: IdOptions<Item>,
-    pub lu_options: ExtractOptions<Item>,
-    pub extract_db_options: ExtractOptions<Item>,
-    pub min_rank: usize,
-    pub symmetric: bool,
-    pub min_level: usize,
-    pub rank_picking: RankPicking,
-    pub num_threads: usize,
-    pub flush_factors: bool,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(bound = "Real<Item>: Deserialize<'de>")]
-pub struct RsrsArgs<Item: RlstScalar> {
-    oversampling: usize,
-    oversampling_diag_blocks: usize,
-    min_num_samples: usize,
-    initial_num_samples: usize,
-    shift: Shift,
-    null_method: NullMethod,
-    qr_method: RankRevealingQrType<Real<Item>>,
-    near_block_extraction_method: BlockExtractionMethod,
-    diag_block_extraction_method: BlockExtractionMethod,
-    lu_pivot_method: PivotMethod,
-    diag_pivot_method: PivotMethod,
-    tol_null: Real<Item>,
-    tol_id: Real<Item>,
-    tol_ext_near: Real<Item>,
-    tol_diag_ext: Real<Item>,
-    min_rank: usize,
-    min_level: usize,
-    symmetric: bool,
-    rank_picking: RankPicking,
-    fact_type: FactType,
-    save_samples: bool,
-    num_threads: usize,
-    flush_factors: bool,
-}
-
-impl<Item> RsrsArgs<Item>
-where
-    Item: RlstScalar,
-{
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        oversampling: usize,
-        oversampling_diag_blocks: usize,
-        min_num_samples: usize,
-        initial_num_samples: usize,
-        shift: Shift,
-        null_method: NullMethod,
-        qr_method: RankRevealingQrType<Real<Item>>,
-        near_block_extraction_method: BlockExtractionMethod,
-        diag_block_extraction_method: BlockExtractionMethod,
-        lu_pivot_method: PivotMethod,
-        diag_pivot_method: PivotMethod,
-        tol_null: Real<Item>,
-        tol_id: Real<Item>,
-        tol_ext_near: Real<Item>,
-        tol_diag_ext: Real<Item>,
-        min_rank: usize,
-        min_level: usize,
-        symmetric: bool,
-        rank_picking: RankPicking,
-        fact_type: FactType,
-        save_samples: bool,
-        num_threads: usize,
-        flush_factors: bool,
-    ) -> Self {
-        Self {
-            oversampling,
-            oversampling_diag_blocks,
-            min_num_samples,
-            initial_num_samples,
-            shift,
-            null_method,
-            qr_method,
-            near_block_extraction_method,
-            diag_block_extraction_method,
-            lu_pivot_method,
-            diag_pivot_method,
-            tol_null,
-            tol_id,
-            tol_ext_near,
-            tol_diag_ext,
-            min_rank,
-            min_level,
-            symmetric,
-            rank_picking,
-            fact_type,
-            save_samples,
-            num_threads,
-            flush_factors,
-        }
-    }
-}
-
-impl<Item: RlstScalar + std::fmt::Display> RsrsOptions<Item> {
-    pub fn new(args: Option<RsrsArgs<Item>>) -> Self {
-        let args = match args {
-            Some(input) => input,
-            None => RsrsArgs::new(
-                8,
-                16,
-                0,
-                420,
-                Shift::False,
-                NullMethod::Projection,
-                RankRevealingQrType::RRQR,
-                BlockExtractionMethod::LuLstSq,
-                BlockExtractionMethod::LuLstSq,
-                PivotMethod::Lu(1e-10),
-                PivotMethod::Lu(0.0),
-                Item::real(1e-10),
-                Item::real(1e-2),
-                Item::real(1e-10),
-                Item::real(1e-10),
-                4,
-                1,
-                false,
-                RankPicking::Min,
-                FactType::Joint,
-                false,
-                num_cpus::get(),
-                false,
-            ),
-        };
-
-        let min_rank = if args.tol_id > num::One::one() {
-            let k = num::ToPrimitive::to_usize(&args.tol_id).unwrap();
-            println!("For tolerances > 1, ID will use this as a fixed rank instead. This fixed rank is: {k}");
-
-            if k <= args.min_rank {
-                k
-            } else {
-                args.min_rank
-            }
-        } else {
-            args.min_rank
-        };
-
-        Self {
-            sketching: SketchingOptions {
-                oversampling: args.oversampling,
-                oversampling_diag_blocks: args.oversampling_diag_blocks,
-                min_num_samples: args.min_num_samples,
-                initial_num_samples: args.initial_num_samples,
-                shift: args.shift,
-                save_samples: args.save_samples,
-            },
-            id_options: IdOptions {
-                null_method: args.null_method,
-                qr_method: args.qr_method,
-                tol_null: args.tol_null,
-                tol_id: args.tol_id,
-            },
-            lu_options: ExtractOptions {
-                block_extraction_method: args.near_block_extraction_method,
-                pivot_method: args.lu_pivot_method,
-                tol_lstsq: args.tol_ext_near,
-            },
-            extract_db_options: ExtractOptions {
-                block_extraction_method: args.diag_block_extraction_method,
-                pivot_method: args.diag_pivot_method,
-                tol_lstsq: args.tol_diag_ext,
-            },
-            fact_type: args.fact_type,
-            min_rank,
-            min_level: args.min_level,
-            symmetric: args.symmetric,
-            rank_picking: args.rank_picking,
-            num_threads: args.num_threads,
-            flush_factors: args.flush_factors,
-        }
-    }
-
-    pub fn to_identifier(&self) -> String {
-        let mut id = String::from("rsrs");
-
-        write!(
-            &mut id,
-            "_null_{:?}_toln_{:e}",
-            self.id_options.null_method, self.id_options.tol_null,
-        )
-        .unwrap();
-
-        match self.sketching.shift {
-            Shift::True(alpha) => write!(
-                &mut id,
-                "_os_{os}_osdiag_{osdiag}_initsam_{init}_shiftd_{alpha:.e}",
-                os = self.sketching.oversampling,
-                osdiag = self.sketching.oversampling_diag_blocks,
-                init = self.sketching.initial_num_samples,
-                alpha = alpha
-            )
-            .unwrap(),
-            Shift::False => write!(
-                &mut id,
-                "_os_{os}_osdiag_{osdiag}_initsam_{init}",
-                os = self.sketching.oversampling,
-                osdiag = self.sketching.oversampling_diag_blocks,
-                init = self.sketching.initial_num_samples
-            )
-            .unwrap(),
-        };
-
-        match self.id_options.qr_method{
-            RankRevealingQrType::RRQR => write!(
-            &mut id,
-            "_mrnk_{}_mlvl_{}_herm_{}_rpick_{:?}_next_{:?}_tolextn_{:e}_db_ext_{:?}_tol_lstsq_{:e}_rrqr",
-            self.min_rank,
-            self.min_level,
-            self.symmetric,
-            self.rank_picking,
-            self.lu_options.block_extraction_method,
-            self.lu_options.tol_lstsq,
-            self.extract_db_options.block_extraction_method,
-            self.extract_db_options.tol_lstsq
-        )
-        .unwrap(),
-            RankRevealingQrType::SRRQR(f) => write!(
-            &mut id,
-            "_mrnk_{}_mlvl_{}_herm_{}_rpick_{:?}_next_{:?}_tolextn_{:e}_db_ext_{:?}_tol_lstsq_{:e}_srrqr_{:e}",
-            self.min_rank,
-            self.min_level,
-            self.symmetric,
-            self.rank_picking,
-            self.lu_options.block_extraction_method,
-            self.lu_options.tol_lstsq,
-            self.extract_db_options.block_extraction_method,
-            self.extract_db_options.tol_lstsq,
-            f
-        )
-        .unwrap(),
-        };
-
-        id
-    }
-}
-
-type Real<T> = <T as rlst::RlstScalar>::Real;
 
 fn oversample<Item: RlstScalar>(
     samples: usize,
@@ -945,7 +638,7 @@ where
         self.stats.tot_id_time += id_step_duration.as_millis();
         println!("ID step in {id_step_duration:?}");
         match &mut rsrs_factors.id_factors {
-            LevelIdFactors::Single(level_factors) => {
+            MultiLevelIdFactors::Single(level_factors) => {
                 level_factors[level_it] = id_factors_res;
                 let start_id_update: Instant = Instant::now();
                 let update_type = UpdateType::Id(&level_factors[level_it]);
@@ -958,7 +651,7 @@ where
 
                 println!("ID updated in {update_id_time:?}\n");
             }
-            LevelIdFactors::Batched(_) => todo!(),
+            MultiLevelIdFactors::Batched(_) => todo!(),
         }
         //rsrs_factors.id_factors[level_it] = id_factors_res;
 
@@ -1215,8 +908,8 @@ where
 
                 rsrs_factors.lu_factors[level_it].push(lu_batch);
                 match &mut rsrs_factors.id_factors {
-                    LevelIdFactors::Single(_) => todo!(),
-                    LevelIdFactors::Batched(level_factors) => {
+                    MultiLevelIdFactors::Single(_) => todo!(),
+                    MultiLevelIdFactors::Batched(level_factors) => {
                         level_factors[level_it].push(id_batch)
                     }
                 }
