+++ src/rsrs/rsrs_factors/mod.rs
@@ -1,5 +1,6 @@
+//! This crate handles the factors used in RSRS. It implements multiplication and inversion when needed
+//!
 pub mod base_factors;
 pub mod commutative_factors;
 pub mod null_and_extract;
 pub mod rsrs_operator;
-pub mod statistics;
